Move all page specific metadata generation to the new PageMetadataBag
